+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 * **2014-04-01**: Refactored the RoutingBundle to provide all routing
   features needed by SimpleCmsBundle.
+  * If you where only using the SimpleCmsBundle you now need to activate
+    the dynamic router. SimpleCmsBundle automatically does this for you.
   * The configuration for document to route/template now all happens under
     cmf_routing.dynamic and the route enhancers also apply to simplecms Pages.
     You can configure additional base paths where to look for routes in the